\name{new_hist}
\alias{new_hist}
\title{Wrapper for Histogram function}
\usage{
new_hist(x, ...)
}
\arguments{
  \item{x}{input vector to compute histogram}

  \item{...}{additional arguments passed to
  \code{\link{hist}}}
}
\description{
This function allows plotting of standardised histograms
with a standardisation by bin (rather than unit)
}
\keyword{utilities}

